commands dir to encapsulate cli logic
